requirements.txt
appdirs==1.4.3
Django==1.10.6
elasticsearch==5.3.0
kafka-python==1.3.3
packaging==16.8
pyparsing==2.2.0
six==1.10.0
urllib3==1.20
